requirements.txt
marimo
pandas
numpy
evaluate
matplotlib
pillow
